+++ pkg/providers/domainownership/testdata/basic_multiple_validation_scopes/import.sh
@@ -1,0 +1,3 @@
+terraform init
+terraform import akamai_property_domainownership_validation.example-com "example.com:DOMAIN,example.com:HOST"
+terraform import akamai_property_domainownership_domains.example-com "example.com:DOMAIN,example.com:HOST"