+++ pkg/providers/domainownership/testdata/basic_one_not_validated/import.sh
@@ -1,0 +1,3 @@
+terraform init
+terraform import akamai_property_domainownership_validation.example-com "example.com:DOMAIN"
+terraform import akamai_property_domainownership_domains.example-com "example.com:DOMAIN,sub.example.com:HOST"