+++ pkg/providers/domainownership/testdata/custom_edgerc_path/variables.tf
@@ -1,0 +1,9 @@
+variable "edgerc_path" {
+  type    = string
+  default = "/path/to/edgerc"
+}
+
+variable "config_section" {
+  type    = string
+  default = "default"
+}